Remove deprecated APIs from `@astrojs/markdown-remark` (#10629)
