+++ .changeset/khaki-otters-return.md
@@ -1,0 +1,5 @@
+---
+"@astrojs/markdown-remark": major
+---
+
+Removes deprecated APIs including `remarkShiki`, `remarkPrism`, `replaceCssVariables` and several unused types
